requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pika
oslo.config>=1.6.0 # Apache-2.0
oslo.utils>=1.2.0 # Apache-2.0
pbr>=0.6,!=0.7,<1.0
python-neutronclient>=2.3.6,<3
python-keystoneclient>=0.10.0
python-novaclient>=2.18.0,<=2.20.0
